fix location of test
